cloud‐run‐slicer/requirements.txt
flask
google-cloud-storage
google-cloud-firestore
imageio-ffmpeg
gunicorn
